silent two compile warnings

+++ src/data.cc
@@ -1057,7 +1057,7 @@ static inline void pinba_update_tag2_report_add(int request_id, pinba_tag_report
 		word1 = (pinba_word *)timer->tag_values[tag1_pos];
 		word2 = (pinba_word *)timer->tag_values[tag2_pos];
 
-		memcpy_static(index_val, record->data.script_name, record->data.script_name_len, index_len);
+		memcpy_static(index_val, record->data.script_name, (int)record->data.script_name_len, index_len);
 		memcat_static(index_val, index_len, "|", 1, index_len);
 		memcat_static(index_val, index_len, word1->str, word1->len, index_len);
 		memcat_static(index_val, index_len, "|", 1, index_len);
@@ -1136,7 +1136,7 @@ static inline void pinba_update_tag2_report_delete(int request_id, pinba_tag_rep
 		word1 = (pinba_word *)timer->tag_values[tag1_pos];
 		word2 = (pinba_word *)timer->tag_values[tag2_pos];
 
-		memcpy_static(index_val, record->data.script_name, record->data.script_name_len, index_len);
+		memcpy_static(index_val, record->data.script_name, (int)record->data.script_name_len, index_len);
 		memcat_static(index_val, index_len, "|", 1, index_len);
 		memcat_static(index_val, index_len, word1->str, word1->len, index_len);
 		memcat_static(index_val, index_len, "|", 1, index_len);
